ObjectListの優先順位を設定可能に

+++ CatchTheBall/Classes/Object.h
@@ -10,10 +10,10 @@ public:
 
 	void Priority(int order);
 	int Priority();
-	void PositionX(int posX);
-	int PositionX();
-	void PositionY(int posY);
-	int PositionY();
+	void PositionX(float posX);
+	float PositionX();
+	void PositionY(float posY);
+	float PositionY();
 	void Rotation(double degree);
 	double Rotation();
 	void Scale(double rate);
@@ -26,7 +26,7 @@ protected:
 private:
 	int _handle;		//•`‰æ‰æ‘œ
 	int _priority;		//—Dæ‡ˆÊ
-	int _x, _y;			//À•W
+	float _x, _y;			//À•W
 	double _rotation;	//‰ñ“]
 	double _scale;		//”{—¦
 };
